+++ apps/api/v2/src/ee/calendars/calendars.module.ts
@@ -3,6 +3,7 @@ import { BookingsRepository_2024_08_13 } from "@/ee/bookings/2024-08-13/bookings
 import { CalendarsRepository } from "@/ee/calendars/calendars.repository";
 import { CalendarsController } from "@/ee/calendars/controllers/calendars.controller";
 import { AppleCalendarService } from "@/ee/calendars/services/apple-calendar.service";
+import { CalendarsCacheService } from "@/ee/calendars/services/calendars-cache.service";
 import { CalendarsService } from "@/ee/calendars/services/calendars.service";
 import { GoogleCalendarService } from "@/ee/calendars/services/gcal.service";
 import { IcsFeedService } from "@/ee/calendars/services/ics-feed.service";
@@ -21,6 +22,7 @@ import { Module } from "@nestjs/common";
   providers: [
     CredentialsRepository,
     CalendarsService,
+    CalendarsCacheService,
     OutlookService,
     GoogleCalendarService,
     AppleCalendarService,
